Fix detection of partial stages
